hanzo-learn: a bounded draw is as wide as its bound
`Twister::below` took a `u32`, and every caller reached it through `i as u32`.
Above 2^32 that is not a refusal, it is a silent truncation: a shuffle of
2^32 + 5 rows drew its positions from `0..=4`, so `subsample((1<<32)+5, 6)`
returned four indices in the first five rows and no error anywhere.

The bound is now a `u64` drawn the way numpy's `random_interval` draws one,
which is two branches and not one. numpy spends a single 32-bit word while the
bound fits in a word and a 64-bit pair — high word first — once it does not.
The width is therefore a function of the BOUND, so an implementation carrying
only one branch is not a simpler version of this: from the first bound past
u32::MAX it stands somewhere else in the stream and every later draw disagrees.

Verified against numpy 2.5.1: the boundary is `most <= u32::MAX`, u32::MAX
itself still costs one word, and the pair is high word first on every seed
tried. The draws that used to collapse to `0..=4` now match
`RandomState(0).randint(0, (1<<32)+5)` exactly.

+++ hanzo-learn/src/isolation.rs
@@ -145,7 +145,7 @@ fn split(
     if rows.len() <= 1 || depth >= limit {
         return leaf(nodes);
     }
-    let feature = rng.below(x.p() as u32 - 1) as usize;
+    let feature = rng.below(x.p() as u64 - 1) as usize;
     let (low, high) = rows.iter().fold((f64::MAX, f64::MIN), |(lo, hi), &i| {
         let v = x.at(i, feature);
         (lo.min(v), hi.max(v))
